Refactor alerts system: fetch all alerts including read status, enhance alert display logic, and implement loading state
